refactor: override proxy behaviour, generalise client redirect uri
The client should not be made aware of its environment as much as
possible. However, when starting the OIDC flow to log the user in,
the client has to manipulate window.location.href. This emits a request
using an http Accept header that, due to [^1], induces react-scripts to
handle the request, rather than proxy it to the backend as desired.

To resolve this, override the default proxy behaviour by using
`src/setupProxy.js`, proxying _all_ requests to `/api/v1/auth` to the
backend.

- `npm install http-proxy-middleware`
- Provide a `src/setupProxy.js` which overrides default proxy behaviour
  so that we dutifully route auth requests to intended destination
- Remove references to `NODE_ENV` in `client/src/api/Sgid`

References

[^1]: https://create-react-app.dev/docs/proxying-api-requests-in-development/